requirements.txt
azure-ai-projects==2.0.0b1
azure-identity
openai
python-dotenv
fastapi
uvicorn[standard]
jinja2
python-multipart
pytest
pytest-asyncio
